Refactors base Isochrones behaviour out to base class Dijkstras (#2182)
* refactor loki interface to take full costing object

* refactor initialization into template so it works with any edge label type

* move time setting out into its own shared function

* move time setting into time dependent initialization

* refactor forward expansion to bdedgelables

* starting point for sucking out basic dijkstras with time dependence

* missed another conflict

* dont compiile this class its not ready  yet

* Decreases number of BDEdgeLabel constructors

* Dijkras compise

* First stab at virtual function for Isochrone customization

* Isochrones: Switch to Dijkra::ExpandForward

* Isochrones: Removes ExpandReverse and migrates Compute to use Dijkras

* Isochrones: Migrates ComputeReverse to use Dijkras

* Isochrones: Migrates ComputeMultiModal to use Dijkras

* Isochrones: ExpandForwardMM to use Dijkras

* Isochrones: Removes Initialize and SetTime no longer needed

* Isochrones: Removes SetOrigin/Destination no longer needed

* minor(Dijkstras): Improves naming of functions

* Resolves added transition_cost from previous merge from master

* change test around a bit to prepare for more difficult testing

* messing around with tests

* test not passing again after refactors to remove potential regressions..

* more testing

* more testing

* more debugging

* redundant

* constify

* superfluous

* all isochrone tests pass

* allow for MM to skip some labels
